BaseIntegrationSpec for all reusable tests
